+++ performance_test/src/communication_abstractions/connext_dds_micro_communicator.hpp
@@ -339,9 +339,8 @@ public:
     }
   }
 
-  std::vector<ReceivedMsgStats> update_subscription() override
+  void update_subscription(MessageReceivedListener & listener) override
   {
-    std::vector<ReceivedMsgStats> stats;
     DDS_Duration_t wait_timeout = {15, 0};
     m_waitset.wait(m_condition_seq, wait_timeout);
 
@@ -354,7 +353,7 @@ public:
       for (decltype(m_data_seq.length()) j = 0; j < m_data_seq.length(); ++j) {
         const auto & data = m_data_seq[j];
         if (m_sample_info_seq[j].valid_data) {
-          stats.emplace_back(
+          listener.on_message_received(
             data.time,
             received_time,
             data.id,
@@ -367,8 +366,6 @@ public:
         m_data_seq,
         m_sample_info_seq);
     }
-
-    return stats;
   }
 
 private: